backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
llama-cpp-python==0.3.0
pydantic==2.5.0
